requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-core
langgraph
aiohttp
motor
pymongo
python-dotenv
python-multipart
jinja2
websockets
pydantic
pydantic-settings
httpx
geopy
pywhatkit